filter command works lfg
