+++ src/message.rs
@@ -1,14 +1,5 @@
 use crate::{display_mode::DisplayMode, error::TGVError, region::Region, strand::Strand};
-use nom::{
-    branch::alt,
-    bytes::complete::{tag, tag_no_case},
-    character::complete::{alpha1, anychar, char, digit1, multispace0, usize},
-    combinator::{map, opt, value},
-    error::ErrorKind,
-    multi::{self, many0, separated_list0},
-    sequence::{delimited, pair, preceded, separated_pair, terminated},
-    IResult, Parser,
-};
+
 use std::str::FromStr;
 use strum::Display;
 
@@ -59,6 +50,8 @@ pub enum StateMessage {
         mouse_released_y: u16,
     },
 
+    AlignmentChange(Vec<AlignmentDisplayOption>),
+
     Quit,
 }
 
@@ -91,42 +84,13 @@ pub enum DataMessage {
     RequiresCytobands(usize),
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Eq, PartialEq)]
 pub enum AlignmentDisplayOption {
     Filter(AlignmentFilter),
 
     Sort(AlignmentSort),
 }
 
-fn parse_display_options(input: &str) -> IResult<&str, Vec<AlignmentDisplayOption>> {
-    many0(alt((parse_filter, parse_sort))).parse(input)
-}
-
-fn parse_filter(input: &str) -> IResult<&str, AlignmentDisplayOption> {
-    delimited(
-        preceded(
-            multispace0,
-            alt((tag_no_case("FILTER"), tag_no_case("WHERE"))),
-        ),
-        node_filter,
-        multispace0,
-    )
-    .parse(input)
-    .map(|(input, filter)| (input, AlignmentDisplayOption::Filter(filter)))
-}
-
-fn parse_sort(input: &str) -> IResult<&str, AlignmentDisplayOption> {
-    delimited(
-        preceded(
-            multispace0,
-            alt((tag_no_case("WHERE"), tag_no_case("ORDER BY"))),
-        ),
-        parse_sort_expression,
-        multispace0,
-    )
-    .parse(input)
-    .map(|(input, filter)| (input, AlignmentDisplayOption::Sort(filter)))
-}
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub enum AlignmentFilter {
     Default,
@@ -175,48 +139,6 @@ pub enum AlignmentFilter {
     Not(Box<AlignmentFilter>),
     And(Box<AlignmentFilter>, Box<AlignmentFilter>),
     Or(Box<AlignmentFilter>, Box<AlignmentFilter>),
-}
-
-fn node_base_filter(input: &str) -> IResult<&str, AlignmentFilter> {
-    use nom::error::ParseError;
-    let (input, (position, base)) = preceded(
-        tag_no_case("BASE"),
-        separated_pair(
-            parse_optional_parenthesis,
-            delimited(multispace0, tag("="), multispace0),
-            alt((
-                tag_no_case("A"),
-                tag_no_case("T"),
-                tag_no_case("C"),
-                tag_no_case("G"),
-                tag_no_case("N"),
-                tag_no_case("SOFTCLIP"),
-            )),
-        ),
-    )
-    .parse(input)?;
-
-    let is_softclip = if base.to_lowercase() == "softclip" {
-        true
-    } else {
-        false
-    };
-
-    let filter = match (position, is_softclip) {
-        (None, true) | (Some(None), true) => AlignmentFilter::BaseAtCurrentPositionSoftClip,
-        (Some(Some(position)), true) => AlignmentFilter::BaseSoftclip(position),
-        (None, false) | (Some(None), false) => {
-            AlignmentFilter::BaseAtCurrentPosition(base.chars().next().unwrap())
-        }
-        (Some(Some(position)), false) => {
-            AlignmentFilter::Base(position, base.chars().next().unwrap())
-        }
-    };
-
-    Ok((input, filter))
-}
-fn node_filter(input: &str) -> IResult<&str, AlignmentFilter> {
-    delimited(multispace0, alt((node_base_filter,)), multispace0).parse(input)
 }
 
 impl AlignmentFilter {
@@ -333,200 +255,6 @@ impl AlignmentSort {
             Self::Default => self,
             Self::Reverse(value) => *value,
             _ => Self::Reverse(Box::new(self)),
-        }
-    }
-}
-
-fn parse_optional_parenthesis(input: &str) -> IResult<&str, Option<Option<usize>>> {
-    opt(delimited(tag("("), opt(usize), tag(")"))).parse(input)
-}
-
-// Parse STRAND with optional number in parentheses
-fn strand_sort_unit(input: &str) -> IResult<&str, AlignmentSort> {
-    let (input, _) = tag_no_case("STRAND")(input)?;
-    let (input, digit) = parse_optional_parenthesis(input)?;
-
-    match digit {
-        Some(Some(position)) => Ok((input, AlignmentSort::StrandAt(position))),
-        _ => Ok((input, AlignmentSort::StrandAtCurrentBase)),
-    }
-}
-
-// Parse STRAND with optional number in parentheses
-fn base_sort_unit(input: &str) -> IResult<&str, AlignmentSort> {
-    let (input, _) = tag_no_case("BASE")(input)?;
-    let (input, digit) = parse_optional_parenthesis(input)?;
-
-    match digit {
-        Some(Some(position)) => Ok((input, AlignmentSort::BaseAt(position))),
-        _ => Ok((input, AlignmentSort::BaseAtCurrentPosition)),
-    }
-}
-
-// Parse basic sort options
-fn sort_unit(input: &str) -> IResult<&str, AlignmentSort> {
-    use nom::Parser;
-
-    alt((
-        base_sort_unit,
-        strand_sort_unit,
-        value(AlignmentSort::Start, tag_no_case("START")),
-        value(AlignmentSort::MappingQuality, tag_no_case("MAPQ")),
-        value(AlignmentSort::Sample, tag_no_case("SAMPLE")),
-        value(AlignmentSort::ReadGroup, tag_no_case("READGROUP")),
-        value(AlignmentSort::ReadOrder, tag_no_case("READORDER")),
-        value(AlignmentSort::ReadName, tag_no_case("READNAME")),
-        value(AlignmentSort::AlignedReadLength, tag_no_case("LENGTH")),
-        value(AlignmentSort::InsertSize, tag_no_case("INSERTSIZE")),
-        value(AlignmentSort::ChromosomeOfMate, tag_no_case("MATECONTIG")),
-        value(AlignmentSort::Tag, tag_no_case("TAG")),
-    ))
-    .parse(input)
-}
-
-// Parse a single sort term (basic sort + optional DESC/DEC)
-fn sort_and_direction(input: &str) -> IResult<&str, AlignmentSort> {
-    let (input, basic_sort) = terminated(sort_unit, multispace0).parse(input)?;
-    let (input, desc_opt) = opt(alt((tag_no_case("DESC"), tag_no_case("ASC")))).parse(input)?;
-
-    match desc_opt {
-        Some(desc) => {
-            if desc.to_ascii_lowercase() == String::from("desc") {
-                Ok((input, basic_sort.reverse()))
-            } else {
-                Ok((input, basic_sort))
-            }
-        }
-        _ => Ok((input, basic_sort)),
-    }
-}
-
-// Parse the complete sort expression
-fn parse_sort_expression(input: &str) -> IResult<&str, AlignmentSort> {
-    let (input, sorts) = delimited(
-        multispace0,
-        separated_list0(
-            delimited(multispace0, char(','), multispace0),
-            sort_and_direction,
-        ),
-        multispace0,
-    )
-    .parse(input)?;
-
-    let result = sorts
-        .into_iter()
-        .reduce(|acc, sort| acc.then(sort))
-        .unwrap_or(AlignmentSort::Default);
-
-    Ok((input, result))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rstest::rstest;
-
-    #[rstest]
-    // Test empty strings
-    #[case("", AlignmentSort::Default)]
-    #[case("   ", AlignmentSort::Default)]
-    #[case("BASE", AlignmentSort::BaseAtCurrentPosition)]
-    #[case("base", AlignmentSort::BaseAtCurrentPosition)]
-    #[case("BASE()", AlignmentSort::BaseAtCurrentPosition)]
-    #[case("base()", AlignmentSort::BaseAtCurrentPosition)]
-    #[case("BASE(2)", AlignmentSort::BaseAt(2))]
-    #[case("base(10)", AlignmentSort::BaseAt(10))]
-    // Test STRAND variants
-    #[case("STRAND", AlignmentSort::StrandAtCurrentBase)]
-    #[case("strand", AlignmentSort::StrandAtCurrentBase)]
-    #[case("STRAND()", AlignmentSort::StrandAtCurrentBase)]
-    #[case("strand()", AlignmentSort::StrandAtCurrentBase)]
-    #[case("STRAND(5)", AlignmentSort::StrandAt(5))]
-    // Test simple keywords
-    #[case("START", AlignmentSort::Start)]
-    #[case("MAPQ", AlignmentSort::MappingQuality)]
-    #[case("readname", AlignmentSort::ReadName)]
-    // Test with DESC/DEC
-    #[case(
-        "BASE(2) DESC",
-        AlignmentSort::Reverse(Box::new(AlignmentSort::BaseAt(2)))
-    )]
-    #[case(
-        "BASE desc",
-        AlignmentSort::Reverse(Box::new(AlignmentSort::BaseAtCurrentPosition))
-    )]
-    #[case(
-        "STRAND desc",
-        AlignmentSort::Reverse(Box::new(AlignmentSort::StrandAtCurrentBase))
-    )]
-    // Test comma-separated (Then)
-    #[case(
-        "BASE(2), START",
-        AlignmentSort::Then(Box::new(AlignmentSort::BaseAt(2)), Box::new(AlignmentSort::Start))
-    )]
-    #[case(
-        "BASE, STRAND(3)",
-        AlignmentSort::Then(
-            Box::new(AlignmentSort::BaseAtCurrentPosition),
-            Box::new(AlignmentSort::StrandAt(3))
-        )
-    )]
-    // Test complex combination
-    #[case(
-        "BASE(2) DESC, MAPQ",
-        AlignmentSort::Then(
-            Box::new(AlignmentSort::Reverse(Box::new(AlignmentSort::BaseAt(2)))),
-            Box::new(AlignmentSort::MappingQuality)
-        )
-    )]
-    // Test with extra whitespace
-    #[case(
-        "  BASE(2)  ,  START  ",
-        AlignmentSort::Then(Box::new(AlignmentSort::BaseAt(2)), Box::new(AlignmentSort::Start))
-    )]
-    fn test_parse_alignment_sort(#[case] input: &str, #[case] expected: AlignmentSort) {
-        let (remaining, sort) = parse_sort_expression(input).unwrap();
-        assert!(remaining.len() == 0);
-        assert_eq!(sort, expected);
-        // TODO: no remaining characters
-    }
-
-    #[rstest]
-    #[case("BASE() DEC")]
-    fn test_parse_alignment_sort_errors(#[case] input: &str) {
-        match parse_sort_expression(input) {
-            Ok((input, sort)) => {
-                assert!(input.len() > 0)
-            }
-            Err(_) => {
-                // Ok
-            }
-        }
-    }
-
-    #[rstest]
-    #[case("BASE=A", AlignmentFilter::BaseAtCurrentPosition('A'))]
-    #[case("BASE(123)=A", AlignmentFilter::Base(123, 'A'))]
-    #[case("BASE=softclip", AlignmentFilter::BaseAtCurrentPositionSoftClip)]
-    #[case("BASE(123)=softclip", AlignmentFilter::BaseSoftclip(123))]
-    #[case("BASE(123) = A", AlignmentFilter::Base(123, 'A'))]
-    fn test_parse_alignment_filter(#[case] input: &str, #[case] expected: AlignmentFilter) {
-        let (remaining, filter) = node_filter(input).unwrap();
-
-        assert!(remaining.len() == 0);
-        assert_eq!(filter, expected);
-    }
-
-    #[rstest]
-    #[case("  BASE=DD  ")]
-    fn test_parse_alignment_filter_error(#[case] input: &str) {
-        match parse_sort_expression(input) {
-            Ok((input, sort)) => {
-                assert!(input.len() > 0)
-            }
-            Err(_) => {
-                // Ok
-            }
         }
     }
 }
